pyproject.toml

[build-system]
requires      = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "skylog_sdk"
version = "0.0.1"
description = "skylog alerting sdk"
readme = "README.md"
authors = [{ name = "Vahid Imanian", email = "vahidtwo@gmail.com" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: BEER-WARE LICENSE",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["skylog", "sdk", "log"]
dependencies = [
    'requests==2.27.1',
    'python-decouple==3.6'
]
requires-python = ">=3.9"

[project.optional-dependencies]
dev = ["black", "bumpver", "isort", "pip-tools", "pytest"]

[project.urls]
Homepage = "https://github.com/vahidtwo/skylog-sdk"
